[bug-69628] make ZipArchiveFakeEntry.setMaxEntrySize publicly accessible
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1924595 13f79535-47bb-0310-9956-ffa450edef68

+++ poi-ooxml/src/main/java/org/apache/poi/openxml4j/util/ZipArchiveFakeEntry.java
@@ -80,7 +80,7 @@ public final class ZipArchiveFakeEntry extends ZipArchiveEntry implements Closea
 
             // Grab the de-compressed contents for later
             data = (entrySize == -1) ? IOUtils.toByteArrayWithMaxLength(inp, getMaxEntrySize()) :
-                    IOUtils.toByteArray(inp, (int)entrySize, getMaxEntrySize());
+                    IOUtils.toByteArray(inp, entrySize, getMaxEntrySize());
         }
     }
